+++ tests/image_test.py
@@ -8,3 +8,10 @@ def test_image_get(client: Beaker, hello_world_image_name: str):
     client.image.get(image.id)
     # Get by name.
     client.image.get(image.name)
+
+
+def test_image_url(client: Beaker, hello_world_image_name: str):
+    assert (
+        client.image.url(hello_world_image_name)
+        == "https://beaker.org/im/01FPB7XCX3GHKW5PS9J4623EBN/details"
+    )