+++ CHANGELOG.md
@@ -10,7 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - `Beaker.dataset.fetch()` and `Beaker.dataset.stream_file()` now verify the digest of the downloaded bytes
-  against the expected digest. A `ChecksumFailedError` is raised if they don't match.
+  against the expected digest by default. A `ChecksumFailedError` is raised if they don't match.
+  You can skip validating the checksum by passing `validate_checksum=False`.
 
 ## [v0.13.0](https://github.com/allenai/beaker-py/releases/tag/v0.13.0) - 2022-04-27